refactor(xaml): move AncestorSource to separate file
Extracted AncestorSource from SyncActivityTable.xaml.cs to AncestorSource.cs and updated XAML namespaces accordingly. No functional changes, improves code organization and maintainability.
