+++ src/gui4/windows/kDrive client/kDrive client/CustomControls/SyncActivityTable.xaml.cs
@@ -11,7 +11,6 @@ using System;
 using System.Collections.ObjectModel;
 using System.ComponentModel;
 using System.IO;
-using System.Reactive;
 using System.Reactive.Linq;
 using System.Threading;
 using System.Threading.Tasks;
@@ -22,58 +21,6 @@ using Windows.ApplicationModel.DataTransfer;
 
 namespace Infomaniak.kDrive.CustomControls
 {
-    public static class AncestorSource
-    {
-        public static readonly DependencyProperty AncestorTypeProperty =
-            DependencyProperty.RegisterAttached(
-                "AncestorType",
-                typeof(Type),
-                typeof(AncestorSource),
-                new PropertyMetadata(default(Type), OnAncestorTypeChanged)
-        );
-
-        public static void SetAncestorType(FrameworkElement element, Type value) =>
-            element.SetValue(AncestorTypeProperty, value);
-
-        public static Type GetAncestorType(FrameworkElement element) =>
-            (Type)element.GetValue(AncestorTypeProperty);
-
-        private static void OnAncestorTypeChanged(DependencyObject d, DependencyPropertyChangedEventArgs e)
-        {
-            FrameworkElement target = (FrameworkElement)d;
-            if (target.IsLoaded)
-                SetDataContext(target);
-            else
-                target.Loaded += OnTargetLoaded;
-        }
-
-        private static void OnTargetLoaded(object sender, RoutedEventArgs e)
-        {
-            FrameworkElement target = (FrameworkElement)sender;
-            target.Loaded -= OnTargetLoaded;
-            SetDataContext(target);
-        }
-
-        private static void SetDataContext(FrameworkElement target)
-        {
-            Type ancestorType = GetAncestorType(target);
-            if (ancestorType != null)
-                target.DataContext = FindParent(target, ancestorType);
-        }
-
-        private static object FindParent(DependencyObject dependencyObject, Type ancestorType)
-        {
-            DependencyObject parent = VisualTreeHelper.GetParent(dependencyObject);
-            if (parent == null)
-                return null;
-
-            if (ancestorType.IsAssignableFrom(parent.GetType()))
-                return parent;
-
-            return FindParent(parent, ancestorType);
-        }
-    }
-
     public sealed partial class SyncActivityTable : UserControl
     {
         private readonly AppModel _viewModel = App.ServiceProvider.GetRequiredService<AppModel>();
